requirements.txt
antlr4-python3-runtime==4.13.2
pytest
pytest-html
pytest-timeout
fastapi==0.109.0
uvicorn==0.27.0
celery==5.3.6
redis==5.0.1
python-multipart==0.0.9